MVP/requirements.txt
quart==0.18.4
redis==4.5.4
PyJWT==2.7.0
pytchat==0.5.5
httpx==0.24.1 